[add/#128] EditEvent model 추가
