+++ app/src/main/java/com/bongtu/baekseo/presentation/recommend/RecommendContract.kt
@@ -4,8 +4,8 @@ import androidx.compose.runtime.Immutable
 import com.bongtu.baekseo.core.common.state.UiState
 import com.bongtu.baekseo.core.common.type.EventType
 import com.bongtu.baekseo.core.common.type.RelationType
+import com.bongtu.baekseo.data.model.event.EditEvent
 import com.bongtu.baekseo.data.model.map.Place
-import com.bongtu.baekseo.presentation.edit.navigation.EditEvent
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
